src/pip-requirements.txt
PyYAML
nose
nose-testconfig
paramiko
pycrypto
requests
scpclient
selenium
pika
restclient
webhdfs
